refactor: type, component, style을 구분하는 주석 추가
